client/requirements.txt
Flask==3.0.0
numpy==1.26.3
Pillow==10.1.0
mss==9.0.1
opencv-python==4.9.0.80
